Move `getSimpleName()` to `TypeName`
